Update dependencies and CI infrastructure
- Update kindlings to 0.1.1, multiarch-scala (and curl) to 0.1.1,
  sge-native-components to 0.1.0, Scala.js to 1.21.0
- TextFormatter: platform-specific MessageFormat via JdkPanama (no
  reflection on JS/Native, graceful fallback to simpleFormat)
- Remove manual curl+extract of native provider JARs from CI — multiarch-
  core handles extraction as transitive dependencies
- Fix IBLBuilder XXX comments, PaletteReducer null-cast to empty array
- Fix int2float implicit conversion in AnimationController, stale @nowarn
  in LittleEndianInputStream

+++ sge/src/main/scala/sge/graphics/g3d/utils/AnimationController.scala
@@ -390,7 +390,7 @@ object AnimationController {
           if (loopCount > 0) loopCount -= 1
           if (loopCount != 0) listener.foreach(_.onLoop(this))
           if (loopCount == 0) {
-            val result = duration * ((loops - 1) - i) + (if (diff < Seconds.zero) duration - time else time)
+            val result = duration * ((loops - 1) - i).toFloat + (if (diff < Seconds.zero) duration - time else time)
             time = if (diff < Seconds.zero) Seconds.zero else duration
             listener.foreach(_.onEnd(this))
             break(result)
